Time: 5 ms (36.77%), Space: 56.7 MB (24.65%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -1,10 +1,10 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
-        int sum = 0;
+        double sum = 0;
         for (int i = 0; i < k; i++) {
             sum += nums[i];
         }
-        int maxSum = sum;
+        double maxSum = sum;
         
         for (int i = k; i < nums.length; i++) {
             sum -= nums[i - k];
@@ -12,6 +12,6 @@ class Solution {
             maxSum = Math.max(maxSum, sum);
         }
 
-        return (double) maxSum / (double) k;
+        return maxSum / k;
     }
 }
